+++ internal/pkg/jwt/jwt.go
@@ -13,11 +13,11 @@ import (
 // CtxKeyJwtUserId get uid from ctx
 var CtxKeyJwtUserId = "jwtUserId"
 
-func GetUidFromCtx(ctx context.Context) int64 {
-	var uid int64
+func GetUidFromCtx(ctx context.Context) uint {
+	var uid uint
 	if jsonUid, ok := ctx.Value(CtxKeyJwtUserId).(json.Number); ok {
 		if int64Uid, err := jsonUid.Int64(); err == nil {
-			uid = int64Uid
+			uid = uint(int64Uid)
 		} else {
 			logx.WithContext(ctx).Errorf("GetUidFromCtx err : %+v", err)
 		}
